refactor: move Server to standalone file
