+++ src/lib.rs
@@ -1,1 +1,8 @@
-pub mod redisapi;
+#[macro_use]
+extern crate quick_error;
+
+mod redisapi;
+mod server;
+
+pub use redisapi::*;
+pub use server::*;
