+++ src/server/errors.rs
@@ -1,0 +1,7 @@
+quick_error! {
+    #[derive(Debug)]
+    pub enum ServerError {
+        RxClosed {}
+        NotStarted {}
+    }
+}
